Restrict values that can be set on a node
This is guard to ensure that complex types, such as map can't be
set with a key inside the dag. By limiting the values to basic
primitive types, arrays of those primitive types, and cids, we can keep
the modification of data inside a node fairly simple - otherwise we
would be tackling traversing inside a node up & down its nested maps.
Setting maps can easily be achieved with SetAsLink, which creates the
map as a new node with a referential cid in the parent.